+++ src-tauri/src/core/mod.rs
@@ -1,3 +1,2 @@
-pub mod kasuri;
 pub mod log;
 pub mod settings;
