+++ src-tauri/src/ui/mod.rs
@@ -14,17 +14,3 @@ pub const EVENT_WINDOW_SHOW: &str = "window-show";
 pub const MENU_ID_EXIT: &str = "exit";
 pub const MENU_ID_RELOAD: &str = "reload";
 pub const MENU_ID_OPEN_LOG_DIR: &str = "open_log_dir";
-
-/// Simplified application data structure used for passing to the UI layer.
-///
-/// This structure contains only the fields needed for displaying
-/// and selecting applications in the user interface.
-#[derive(serde::Serialize)]
-pub struct AppForView {
-    /// Display name of the application
-    pub name: String,
-    /// Unique identifier for the application
-    pub app_id: String,
-    /// Path to the application icon
-    pub icon_path: String,
-}
